plot3D-package {plot3D}	R Documentation

Plotting multi-dimensional data.

Description

Functions for visualising 2-D and 3-D data.

Many of the functions are extensions of R's persp or image function.

Other packages that provide visualisation of 3-D data (and which might be better suited) are: rgl, scatterplot3D, misc3D.

Note

This package is dedicated to Carlo.

Note

Some of the functions based on persp will not work properly for all values of phi (which turns the plots upside-down). This is because an assumption is made as to how the perspective plots are viewed.

Author(s)

Karline Soetaert

References

http://www.rforscience.com/rpackages/visualisation/oceanview/

http://www.rforscience.com/rpackages/visualisation/plot3d/

See Also

Functions that are based on the persp function:

persp3D: an extended version of persp.

ribbon3D: a perspective plot as ribbons.

hist3D: 3-D histograms.

scatter3D, points3D, lines3D: colored points, lines, ... in 3-D.

slice3D, slicecont3D: slices from a full 3-D data set.

isosurf3D: isosurfaces from a full 3-D data set as triangles.

voxel3D: isosurfaces from a full 3-D data set as points.

surf3D, spheresurf3D: 3-D shapes or surfaces.

arrows3D: arrows in 3-D.

segments3D: line segments in 3-D.

polygon3D: 3-D polygons.

box3D, border3D, rect3D: boxes and rectangles in 3-D.

text3D: labels in 3-D.

Functions defined on the image function:

image2D, for an image function to visualise 2-D or 3-D data.

ImageOcean: an image of the ocean's bathymetry.

Other plotting functions:

contour2D, for a contour function to visualise 2-D data and that have a color key.

scatter2D: colored points, lines, ... in 2-D.

text2D, arrows2D, segments2D, rect2D, polygon2D for other 2D functions that have a color key.

Colors and colorkey:

colkey: adds a color legend.

jet.col, ramp.col, gg.col, alpha.col: suitable colors, shade and lighting.

Utility functions:

mesh: to generate rectangular (x, y) or (x, y, z) meshes.

Data sets:

Oxsat: 3-D data set with the ocean's oxygen saturation values.

Hypsometry: 2-D data set with the worlds elevation and ocean's bathymetry.

Examples

# run all examples
## Not run: 
 example(persp3D)  
 example(surf3D)  
 example(slice3D)
 example(scatter3D)
 example(segments3D)
 example(image2D)
 example(image3D)
 example(contour3D)
 example(colkey) 
 example(jet.col) 
 example(perspbox)
 example(mesh)
 example(trans3D)
 example(plot.plist)
 example(ImageOcean)
 example(Oxsat)

## End(Not run)

[Package plot3D version 1.4 Index]